python/examples/spot_cam/requirements.txt
-f ../../../prebuilt

bosdyn-client >= 3.1

Pillow==6.1.0
numpy==1.18.5
aiortc==0.9.28
requests==2.26.0
opencv-python >= 3.4.2.17
pylibsrtp==0.6.8
av==8.0.3
